BaseNode et.al.: protect against overflow, store compacted values as unsigned int

+++ jol-addons/src/main/java/org/openjdk/jol/addons/InitialNode.java
@@ -38,8 +38,8 @@ final class InitialNode extends BaseNode implements HeapWalker.Node {
     }
 
     @Override
-    public int getCount() {
-        return 1; // one InitialNode per heap object
+    public long getCount() {
+        return 1L; // one InitialNode per heap object
     }
 
     public Class<?> getObjectClass() {
